Update test implementation
We were using sys.path[0] as our known output value, but it was
not constant. Hence, it wasn't really a known output value.

Switched to using os.path.dirname() which should be a good enough
for us to equate our function results to, as it does what the
function is doing and if the function is changed in a way such
that it changes its output value, the test will pick it up.