Automatically clear screenshots when a checkpoint occurs.
This is a compromise between the 6.99 and 7.x behavior. In the older
stuff, we would never automatically clear the screenshot, meaning
if a game called renpy.take_screenshot, it could grow stale.

In the newer behavior, we would clear it after each interaction,
even meanigless ones, like an implie with.

This uses renpy.checkpoint to clear the screenshot, which is a proxy
for the game state meaningfully advancing. In addition, this behavior
can be disabled by setting config.auto_clear_screenshot to False.

Fixes #1680.