sl2: Do not update scopes in used screens.
This is a much simpler fix for the problem in a1dd218560d127f1453f3c57d3dd4da92df54063,
which is that the wrong scope will be given in a used screen that happens to be
constant.

The realization is that if a used screen is constant, it can't change - and hence
it doesn't need to be given a new scope at all.