py3: Trigger .items() change differently.
It now triggers on: from __future__ import with_statement, division.

This should be a bit safer, because with_statement is obsolete - even if
it is used by any Python, it can be removed without changing the
semantics of the program. (We can't trigger _just_ on it, as all
old .rpycs are compiled with that flag.)