gl2: Improvements related to properties and LOD.
LOD is now controlled by u_lod_bias.

There is now a dictionary that is passed around through the gl2
renderer, properties. It's intended as the place to store
information that controls the fixed-function portions of
the GL pipeline. Right now, it controls the texture_scaling
mode (used to implement nearest neighbor rendering), and
pixel_perfect rendering, but it can be used for other things
in the future.

The full set of texture_scaling modes are now available.