Add a controller blocklist for the Switch pro controller.
This controller doesn't work without some other software to send
the init to it. Trying to use hidapi to initialize the controller
didn't work, and led to problems with other software that uses
USB. (And fundamentally, I don't think Ren'Py should use direct USB
access.)

So, instead, we detect the pro controller and ignore it.