Fix the interaction between xfill, yfill, and box_wrap.
First off, it makes sense that if a size is being forced on a
displayable, it should fill to fit that size in possible. So
now xsize, ysize, and xysize force xfill and/or yfill in the
appropriate dimensions.

If box_wrap is true, however, fill is turned off for the last
line. This prevents it from being expanded unnaturally.

Fixes #2589, while preserving the fix for #2081.