relative volume management independant of context
Possibly, take this commit as a guide or idea to approach this.

Adds the ability for channels to manage their "tertiary" volume independently of context.
In addition, set this tertiary volume for a context switch (loading).

Tertiary volume is set per-track, such as:
renpy.music.play("music.opus", relative_volume=0.7)  
  or
play music "music.opus" volume 0.7

As opposed to "secondary" volume which is set by channel controls, like
$ renpy.music.set_volume(0.4, delay=0.5, channel="music")

The idea is that things like loading a game can cause sudden differences in 'track' volumes. Since these volumes were stored with the context, and immediately applied via mixing with the 'secondary' volume, there was no way (I could see) to allow smooth volume transitions when  fade-in/fade-out occurred with the context switch,

For instance, if one had a track playing at 0.3 tertiary volume, then loaded a game with 1.0 volume,  the fades would occur, but the channel tertiary volume would immediately jump to 1.0, making the old track play very loudly for some part of the fadeout duration.

"loop_only" in the enqueue function is only set by music.play when the new loop list is exactly the same as the old

I attempted to do this in a way that minimized change to behavior.
Some decisions were made:
If a new list of loop tracks are queued, and the currently playing track is among them, then immediately alter the volume.
Also, added a 'dequeue' to the context-switch interact, when the channel is currently playing a loop. Previously, the channel would 'fadeout' (and so clear the queue and playing track) if the current playing track wasn't in the upcoming filename list,  but if it *was*,  it would NOT dequeue, but still queue the whole list, potentially adding many copies of the filename list to the queue if several loads were done in a row.