Remove Transform.show()/ATLTransform._show()
These methods don't do much in the current Ren'Py. ATL
images are predicted by .visit(), while Transform's child is set
when it's added. The only think it's used for is setting the
default_transform's state, and that was fixed by a direct call
to update_state.()

What's more, _show() has caused problems, first #2834, and now