Limit when a Button propagates transform state.
It now only does it through a reference when it is a reference
with the same name.

This is a compromise behavior, at best - at some point, we need
to switch to a persistent dag rather than what we have now.

Per p547535 on LSF, and also takes into account #2864.