Always run an unfocused handler when a displayable remains shown.
Ren'Py will now always run an `unhovered` action when a displayable (or its
replacement) remains shown, and the focus changes. Previously, the unhovered
action would not run when the loss of focus was caused by showing a second
screen.

This traces back to LSF PM 295318 (and the rest of that thread.)