Add config.default_attributes callback dictionary
This is to support images having a "weak" set of default attributes
applied to them when they are first shown, allowing for more complex use
cases with a minimum of boilerplate in scripts.

While similar in nature to the existing config.adjust_attributes, this
feature only applies to images when they are first shown, and is "weak"
in the sense that it may not countermand attributes from the script.