distribute: Base vc_version on the date.
It's now YYMMDDCC, where CC is the commit of the day starting with 1.

The goal is to make it easy to see how old a given nightly or release
is just by looking at the full version, and it also means it's
not necessary to tag that start of each release.