Ignore errors in visit_all in SceneLists.add.
This only happens when config.developer is not true. It's meant
to fix an issue introduced in 7.4, where a displayables
visit_all method can be called before an interaction begins.

That's a problem with ATL that depends on a variable that's defined
between the show and interaction - something like:

    show atl_that_uses_foo
    $ foo = True
    "This."

The reasoning for visit_all was to allow the state of the ATL to
be taken (say, by having multiple shows in a row) without an
interaction to take place. Since the older games with this
problem can't successfully do this, it's okay to ignore this
error.

Fixes #3229.