Improve duplicatable check to consolidate coercion
Trying to coerce the child arg to Transform.__call__ into a displayable
adds an avoidable call and could lead to redundant duplication (in the
case where the coercion yields a duplicatable displayable).

Instead, use getattr in the duplicatable check to prevent raising and
allow __init__ (which already does arg coercion) to handle it as normal.