Allow get_playing() to return queued audio before channel number is allocated.
This prevents a race condition caused by something being enqueued
but not played before renpy.get_playing() is called, the first
time the channel was used in a session.

This mostly manifested in in renpy.movie_cutscene(), which could
skip the movie if the outcome of the race was wrong.