Wrap comprehensions and generator expressions in lambdas.
This fixes #3274, which is a subtle bug that was actually there in
Python 3, if rarely encountered. The problem is fundamentally
that:

eval("{ i for i in l if i == b }", globals={ "l" : [1, 2, 3]}, locals={ "b" : 2 })

Doesn't work, since the comprehension (non-list comprehensions in
py2, list comprehensions also in py3, generator expressions in
all versions) creates its own scope, and since b isn't defined in
an exterior scope, b is loaded with the LOAD_GLOBAL opcode, which
doesn't check locals.

To fix this, the statement is wrapped in a lambda that is immediately
called with the variables it loads - so the expression above is
turned into:

    (lambda l, b : { i for i in l if i == b })(l, b)

This creates a new scope, which the python compile can find,
and compile the generator expression or comprehension to
pull from.