Only trigger first frame once in fast-forward
When fast-forwarding through ATL, one frame from every interpolation is
shown due to the first_frame flag. To avoid this, only set the flag if
the st rests within the duration of the interpolation. This ensures that
the flag is correctly set, but only for the interpolation that will be
active at the st.