Skip recording deltas for stores that don't change
Currently keys for every store are added to a Rollback object's stores
and delta_ebc dicts, even if they contain no data. When rolling back
these dicts are iterated over, and the empty stores amount to no-ops.
These empty data structures are also serialised/de-serialised by the
save system (albeit efficiently) but yield no benefit.

This change is to skip recording entries for stores that haven't
changed, the resulting log should be functionally equivalent to existing
behaviour, but cut down on the amount of extraneous work to handle them.

As the code processing the log doesn't change or mind stores being
omitted, this change should be completely transparent and both forward
and backward compatible.