Add tinyfiledialogs.
This is a library that's intended to be used to provide os-native
file picking services to the launcher, to replace the various
libraries that are used now.

This is included with Ren'Py because it's not in pypi or in
operating systems, and it's needed for an os-linked build of
Ren'Py to work.