Disable roll forward for call screen.
The main justification for this is that roll forward only preserves
the return value of the screen, or a jump out of it, and doesn't
preserve any other side effects in the screen. This can lead to the
game being in a different state between playthoughs.

A new roll_forward property allows it to be re-enabled for particular
screens.

Alternatives considered were:

* Trying to determine which actions should allow a screen to
  roll forward. This was implemented, but was rejected as a maintenance
  problem, and hard to make understandable in the documentation.

* The reverse of rollback was considered, but it seems hard to make
  work. What should be done if the player changes sone data inside
  the call screen, then rolls forward, causing a different set of
  data to be updated?

Fixes #2563.