Rewrite how relative_volume works internally.
This goes back to the original idea of it, which was to provide a
way of normalizing track volumes. This change makes it so that the
relative volume selected sticks with the track.

This means that, now, if you queue a track with a given relative
volume, that only takes effect when the track is played. Other
changes make relative volume  work better with rollback, and
loads.

Fixes #2654.