Remove __init__ from new Partial class
The underlying functools.partial doesn't have an __init__ method, so the
super call bounces up to object which doesn't take arguments. This is
fine in normal use but presented issues in some pickling scenarios.

Since we can be slightly more efficient without it, and it avoids the
edge cases with pickle, nuke it.