Refactor Curry class into a Partial factory
This allows pickled instances of the old Curry class to deserialise
cleanly into functionally-equivalent Partial instances.

This is done by having it's __new__ method return a Partial instance
created with a placeholder callable (hash), and rebound setstate method
capable of converting the state of the old Curry instance into the
format that functools.partial expects.

Any attempt to create a new instance of the Curry class directly will
result in an instance of Partial, avoiding further propagation of the
deprecated class.

As all of the extra work for compatibility only occurs when attempting
to deserialise Curry instances, we don't impact performance of the new
Partial class, or compromise it's API into the future.