Do not call _unique from _duplicate.
Fixes #3349.

The _unique method should only be called if a displayable will only be
entering the tree of displayables to be rendered once. That
wasn't the case in certain circumstances, when _unique was
being called from _duplicate, which is usually called when a
displayable is added.

This might trade performance (additional duplications) for
correctness. That's something we can accept for now, and
eventually deal with in the future by optimizing when
displayables are duplicated.