Preferences screen redesign.
This moves the preferences screen to a multi-tab design, with
the left 1/3 being a list of tabs and the right 2/3 being the
tab content.

The motivation for this was that the preferences were getting
overfull, and some translations were breaking the layout. By
dividing the screen up, we get a lot more space to play
with.