compat: Reorder globals within a function.
Newer versions of Python 3 enforce the condition that a global
statement has to come before variables are assigned, while older
Python versions didn't enforce this.

Since games rely on this to work, this pass finds all the global
statements inside the function, replaces them with pass statements,
and then creates a new global statement at the start of the function.

This isn't needed for nonlocal, as no games currently use that
statement.