Invert how mark_modal works.
The find_focus method iterates from back to front, and so when
mark modal is called, the focuses behind it (that have already
been given to the callback) are the ones that can't be given focus.

The previous implementation got that wrong - the displayables in
front were the ones that couldn't be given focus, which lead to
random problems.

Fixes #3617.