Scroll action with float amount and delay
I'm not sure if it is a good code but it works pretty well for me. I've just added two features:
`amount` - now it can receive a float number between 0.0 and 1.0. This number will be converted to pixels.
`delay` - scrolling with ease interpolation. It won't be scrolling if the scrolling range equals zero.