py3: Update mobilebuild exception logging
In py3 the print_exc function prints a str, which files opened in binary
mode do not like. Conversely py2 doesn't seem to mind whether the file
is in binary mode or not.

This matches the approach taken in the rest of the file, where various
logs and outputs are also not opened in binary mode any longer.