sl1: Remove the ability to parse SL1 screens.
This had been left in to support potential issues in the change
from SL1 to SL2. That wasn't used, and almost certainly isn't
being used anymore, so the parser can be removed.

SL1 screens can still be loaded if they exist in SL1 files,
as this only affects the parser, not the runtime.

Fixes #3430.