Adjust how autoreload loading and saving works.
This attempts to address #3799, which I'm guessing is caused by
the game being saved when in an error handling context that's
handling a parse error.

What this does is to try to retain the initial save that is created
until the save is loaded successfuly, or alternatively until the
game quits.

By not creating a new save slot for each reload attempt, this
tries to avoid creating a save when Ren'Py is in an unstable
state.