Allow viewport drag to override child focus.
Now, when a a touch moves by a certain number of pixels when a
button inside a viewport is focused, focus and grab are shifted
onto the viewport, which then can be scrolled.

This also changes several things about how viewports process
events - now, if an event can be handled by the viewport, the
viewport takes precendence.