Improvements to viewport dragging.
The biggest change here is that there is a new _draggable field on
Displayable. If True, the displable itself might be draggable, and
so a Viewport won't steal the focus of it.

There are also several changes that make sure the viewport gets and
releases drags as appropriate.

Finally, draggability only enables when the displayable is actually
scrollable in at least on axis.

Fixes #3910.