New behavior regarding modal, pause, and timers.
By default, timers trigger even if uderneath a modal screen
- a screen timer is intended primarily to display data, and not
by default, as a means of control. (For example, the default
notify screen uses a timer to hide itself without game effect.)

Pauses, by default, respect modal screens, as pause is part of
the main game's control flow.

Both timers and pauses take a modal property, which overrides
if a modal screen is respected or not.

Fixes #3935.