Jump to the deferred translate identifier if the translate changes.
This will re-run, and hence re-translate, say statements that have
interact=False, most notably the Say statement that precedes a
menu statement.

This could cause problems if there is a non-idempotent statement
between the say statement and the interaction, like:

    e "This will show dialogue when the screen is shown." interact False

    $ variable += 1

    call screen myscreen

That should be rare enough that a warning will suffice.