web: Process the JS command queue regularly.
This moves the JS command queue processing into the 20hz periodic
event handler. Moving it out of the idle frame handing makes sure
that the command queue runs even if the the idle frame code is
being starved by frequent events or animations.

The js command overhead was measured, and seems to be less than
the browser's cock precision.