Improve numeric keypad keysyms.
Two changes in this one. The first is to add a quartet of new
modifiers (caps/nocaps, num/nonum) to reflect if capslock or
numlock are turned on.

The second is to create a new type of keysym, beggining with KP,
that expands out to the key with our without the numlock modifier.

Starts work on #4030.