Allow side to expand to fit its child on render.
This deals with the case where the the insets caused by the
side displayable might cause the center displayab;e to grow in
one dimension. (For example, Text that has an inset on the right
may be bigger vertically than if there was no inset.)

When this happens, the space allocated to the center displayable
will be allowed to expand until everything fits.

Fixes #4253.