Add Lexer.deferred_error, and support for deferred errors.
A deferred error is a parse error that might or might not manifest,
based on configuration variables that need to be set early, but can
be in any file.

These errors are stored in .rpyc files, and might become real parse
errors, or might be discarded.

This is intended to address #4284.