sl2: Analyze for loops with break/continue for constness.
When a for loop has a break or continue in it, it may not
complete all of its executions. The result of that is that
variables that are updated in the loop are no longer constant
in the screen.

This has break and continue mark the loop as non-constant, and
then the loop re-analyzed to mark the variables in it as
non-constant.