The current screen shouldn't be a HIDE or OLD screen.
There are various actions and values that access displayables
in the current screen. Some of these might be run (for example,
BarValue.get_adjustment()) in a hiding or old(-state-of-transition)
screen.

Almost certainly, those values and actions want the displayables
that are present in the original screen, not the copy, and so
the value is redirected.

Fixes #3998.