Include an example of a proper usage of time_warp in the docs of ImageDissolve (#4326)
Include an example of a proper usage of time_warp in ImageDissolve

Even though time_warp is mentioned three times in the transition classes document, it's never shown how to actually use it. Since the idea of using the _warper module is not intuitive, and it's mentioned only once at the beginning of the section, I thought it would be best to at least include it in one of the examples.