per-sample: Implement volumes.
This implements all four volumes that can be applied to a stream -
mixer, relative_volume, secondary_volume, and fade.

The interpolated volumes (secondary and fade) are now interpolated
using power, rather than magnitude, which leads to perceptually
linear interpolation.

There is a bit of a performance enhancement here - instead of
using decibels as our power units, we use a base-2 power unit
that can be calculated faster.