per-sample: Prevent clicking noises.
There were two causes for these. The first was that if a sound
was queued tight, the fadeout would finish, and both playing and
queued sounds would be removed - meaning nothing would play
until the next time periodic ran. Fadeout now dequeues the
queued sound if the fadeout would finish.

The second was a mistake that looked at the wrong portion of the
stream buffer, with the resulting errors often sounding like
clicks.